Refactor tach for flexibility.
Separates the concepts of notifiers and statistics; selects the
notifier based on the configuration file section included.  The
statistic can be expressed as a class name suitable for a call
to utils.import_class_or_module(), or, for compatibility, by the
old names from tach.decorator (which is now gone).